Extract everything heap-related into a file
Leave in analyzer only the visualization
